// spike/159-kysely-migration-tooling/src/schema.generated.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface Users {
  createdAt: Generated<Timestamp>;
  email: string;
  id: string;
  name: string;
  passwordHash: string | null;
  passwordResetToken: string | null;
  passwordResetTokenExpiresAt: Timestamp | null;
  seed: Generated<number>;
  updatedAt: Generated<Timestamp>;
  username: string;
}

export interface DB {
  users: Users;
}